+++ packages/queue/src/queue/createQueue.ts
@@ -3,12 +3,12 @@ import { QueueRepository } from "./queueRepository";
 import { ConnectionManager } from "../connectionManager";
 import { JobData, QueueOptions } from "../types";
 import { SharedContext } from "@alanszp/shared-context";
+import { RepeatableQueue } from "./repetableQueue";
 
 export function createQueue<Data = JobData>(
   name: string,
   getContext: () => SharedContext,
-  queueOptions?: QueueOptions,
-  addDefaultOptions?: boolean
+  queueOptions?: QueueOptions
 ): Queue<Data> {
   const connection = ConnectionManager.getInstance().getConnection();
   const queue = new Queue<Data>(
@@ -16,8 +16,24 @@ export function createQueue<Data = JobData>(
     name,
     ConnectionManager.getInstance().getServiceName(),
     getContext,
-    queueOptions,
-    addDefaultOptions
+    queueOptions
+  );
+  QueueRepository.Instance.registerQueue(queue);
+  return queue;
+}
+
+export function createRepeatableQueue<Data = JobData>(
+  name: string,
+  getContext: () => SharedContext,
+  queueOptions?: QueueOptions
+): RepeatableQueue<Data> {
+  const connection = ConnectionManager.getInstance().getConnection();
+  const queue = new RepeatableQueue<Data>(
+    connection,
+    name,
+    ConnectionManager.getInstance().getServiceName(),
+    getContext,
+    queueOptions
   );
   QueueRepository.Instance.registerQueue(queue);
   return queue;