+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Upgrade bullmq to 5.x in `@alanszp/queues`
 - Upgrade ioredis to 5.3.2 `@alanszp/queues`: 5.3.2 or higher is required
 - Change `QueueOptions` & `WorkerOptions` types on `@alanszp/queues`: Connection is no longer a prop, because it's already required in the queue and worker constructor.
-- Add `addDefaultOptions` constructor property (optional, boolean, default = true) on `@alanszp/queues`: Defines wether to add or no default queue configuration.
+- Add `RepeatableQueue` class `@alanszp/queues`: for repeatable jobs
 
 ## v13.0.0